requirements.txt
asgiref==3.4.1
cloudinary==1.26.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.5
Django==3.2.8
django-allauth==0.45.0
django-crispy-forms==1.13.0
django-summernote==0.8.11.6
django-taggit==1.5.1
gunicorn==20.1.0
oauthlib==3.1.1
psycopg2==2.9.1
PyJWT==2.2.0
python3-openid==3.2.0
pytz==2021.3
requests-oauthlib==1.3.0
sqlparse==0.4.2
